# pkg/termite/openapi.yaml
openapi: "3.0.3"
info:
  version: v1alpha1
  title: Termite API
  description: |
    Termite is an Ollama-like local inference server for ONNX-based ML models.

    ## What is Termite

    Termite provides local ML inference with an Ollama-compatible API:
    - **Embedding Generation**: Text and multimodal (CLIP) embedding models
    - **Text Chunking**: Semantic chunking with ONNX models or fixed-size fallback
    - **Reranking**: Relevance re-scoring for search results
    - **Future**: NER, classification, and generative model support planned

    Download the latest release at https://antfly.io/docs/downloads

    ## When to Use Termite

    Termite can run standalone or as part of an Antfly cluster:
    - Local ONNX model inference without external API dependencies
    - Ollama-compatible `/api/embed` endpoint for embeddings
    - Semantic text chunking for RAG pipelines
    - Relevance reranking for improved search quality
    - Centralized model serving across distributed nodes
    - Privacy-preserving ML inference (data never leaves your infrastructure)

    ## Features

    ### Embedding Generation
    - **Models**: ONNX models auto-discovered from `embedder_models_dir`
    - **API**: Ollama-compatible `/api/embed` endpoint
    - **Response Formats**: Binary (default), JSON

    ### Multimodal Support (CLIP)
    - **Image Embeddings**: CLIP models for joint text-image embedding space
    - **Input Formats**: Base64 data URIs (`data:image/png;base64,...`) or URLs
...

...

...

...
servers:
...
components:
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message

    # Multimodal Content Types (OpenAI-compatible)
    TextContentPart:
      type: object
      description: Text content for embedding
      required:
        - type
        - text
      properties:
        type:
          type: string
          enum: [text]
        text:
          type: string
          description: Text content to embed
    ImageURL:
      type: object
      description: Image URL or data URI
      required:
        - url
      properties:
        url:
          type: string
          description: URL or data URI (data:image/png;base64,...)
          example: "data:image/png;base64,iVBORw0KGgo..."
    ImageURLContentPart:
      type: object
      description: Image content for embedding (OpenAI-compatible format)
      required:
        - type
        - image_url
      properties:
        type:
          type: string
          enum: [image_url]
        image_url:
          $ref: "#/components/schemas/ImageURL"
    ContentPart:
      description: A content part for multimodal embedding (text or image)
      oneOf:
        - $ref: "#/components/schemas/TextContentPart"
        - $ref: "#/components/schemas/ImageURLContentPart"

    # Embedding Types (Ollama-compatible with multimodal extension)
    EmbedRequest:
      type: object
      required:
        - model
        - input
      properties:
        model:
          type: string
          description: Name of the embedder model from embedder_models_dir
          example: "bge-small-en-v1.5"
        input:
          oneOf:
            - type: string
              description: Single text string (backward compatible)
            - type: array
              items:
                type: string
              description: Array of text strings (backward compatible)
            - type: array
              items:
                $ref: "#/components/schemas/ContentPart"
              description: Array of multimodal content parts (text or images)
          description: |
            Input content to embed. Supports three formats:
            - Single text string: `"hello world"`
            - Array of text strings: `["hello", "world"]`
            - Array of content parts (multimodal): `[{"type": "text", "text": "hello"}, {"type": "image_url", "image_url": {"url": "data:image/png;base64,..."}}]`
          example: ["hello world", "machine learning"]
        truncate:
          type: boolean
          default: true
          description: Truncate input to fit model context length
    EmbedResponse:
      type: object
      example:
        {
          "model": "bge-small-en-v1.5",
          "embeddings": [[0.0123, -0.0456, 0.0789], [0.0234, -0.0567, 0.0890]],
        }
      required:
        - model
        - embeddings
      properties:
        model:
          type: string
          description: Model used for embedding
          example: "bge-small-en-v1.5"
        embeddings:
          type: array
          items:
            type: array
            items:
              type: number
              format: float
          description: Array of embedding vectors (one per input string)

    # Chunking Types - reference existing schemas
    Chunk:
      $ref: "../../../antfly-go/libaf/chunking/openapi.yaml#/components/schemas/Chunk"
    ChunkConfig:
      type: object
      description: |
        Configuration for chunking requests to Termite API.
        This is a simplified config for the HTTP API - differs from the full ChunkerConfig
        which includes provider selection and caching configuration.
      properties:
        model:
          type: string
          description: "The chunking model to use. Either 'fixed' for simple token-based chunking, or a model name from models/chunkers/{name}/."
          default: "fixed"
          example: "fixed"
        target_tokens:
          type: integer
          description: Target number of tokens per chunk
          example: 500
          default: 500
        overlap_tokens:
          type: integer
          description: Number of overlapping tokens between chunks
          example: 50
          default: 50
        separator:
          type: string
          description: Text separator for fixed chunking
          example: "\n\n"
          default: "\n\n"
        max_chunks:
          type: integer
          description: Maximum number of chunks to return
          example: 50
          default: 50
        threshold:
          type: number
          format: float
          description: Confidence threshold for ONNX models (0.0-1.0)
          example: 0.5
          default: 0.5
    ChunkRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Text to chunk
          example: "This is a long document that needs to be split into smaller chunks..."
        config:
          $ref: "#/components/schemas/ChunkConfig"
    ChunkResponse:
      type: object
      required:
        - chunks
        - model
        - cache_hit
      example:
        {
          "chunks":
            [
              {
                "text": "This is the first chunk...",
                "start_offset": 0,
                "end_offset": 100,
                "metadata": {},
              },
              {
                "text": "This is the second chunk...",
                "start_offset": 90,
                "end_offset": 190,
                "metadata": {},
              },
            ],
          "model": "fixed",
          "cache_hit": false,
        }
      properties:
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/Chunk"
          description: Array of text chunks
        model:
          type: string
          description: Chunking model actually used (may differ from requested if fallback occurred)
          example: "fixed"
        cache_hit:
          type: boolean
          description: Whether result was served from cache

    # Reranking Types
    RerankRequest:
      type: object
      required:
        - model
        - query
        - prompts
      properties:
        model:
          type: string
          description: Name of reranking model from `reranker_models_dir`
          example: "bge-reranker-v2-m3"
        query:
          type: string
          description: Search query for relevance scoring
          example: "machine learning applications"
        prompts:
          type: array
          items:
            type: string
          description: |
            Pre-rendered document texts to rerank. The client is responsible for extracting
            and rendering document fields/templates before calling this endpoint.
          example:
            [
              "Introduction to machine learning...",
              "Deep learning fundamentals...",
            ]
    RerankResponse:
      type: object
      required:
        - model
        - scores
      properties:
        model:
          type: string
          description: Name of model used for reranking
        scores:
          type: array
          items:
            type: number
            format: float
          description: Relevance scores (one per prompt, same order as input)

    # Models Types
    ModelsResponse:
      type: object
      required:
        - chunkers
        - rerankers
        - embedders
      properties:
        chunkers:
          type: array
          items:
            type: string
          description: Available chunking models (always includes "fixed")
          example: ["fixed", "chonky-mmbert-small-multilingual-1"]
        rerankers:
          type: array
          items:
            type: string
          description: Available reranking models
          example: ["bge-reranker-v2-m3"]
        embedders:
          type: array
          items:
            type: string
          description: Available embedding models from embedder_models_dir
          example: ["bge-small-en-v1.5", "bge-small-en-v1.5-i8-qt"]
...
paths:
  /embed:
    post:
      summary: Generate embeddings
      description: |
        Generates vector embeddings for input content using local ONNX models.
        This endpoint is compatible with Ollama's `/api/embed` format for text,
        and extends it with OpenAI-compatible multimodal support for CLIP models.

        ## Models

        Models are auto-discovered from `embedder_models_dir` at startup.
        Use the `/api/models` endpoint to list available models.

        - **Text-only models** (e.g., bge-small-en-v1.5): Accept text strings
        - **Multimodal models** (e.g., CLIP): Accept text and images via data URIs

        ## Input Formats

        Three formats are supported:
        - Single text string: `"hello world"`
        - Array of text strings: `["hello", "world"]` (Ollama-compatible)
        - Array of content parts: `[{"type": "text", "text": "..."}, {"type": "image_url", "image_url": {"url": "data:image/png;base64,..."}}]` (OpenAI-compatible)

        ## Caching

        Results are cached in memory for 2 minutes. Concurrent identical requests are deduplicated
        using singleflight to prevent redundant work.

        ## Response Formats

        Supports multiple content types via Accept header:
        - `application/octet-stream`: Binary serialization (default, most efficient)
        - `application/json`: JSON response with model name and embeddings

        ## Examples

        Text embedding (Ollama-compatible):
        ```json
        {
          "model": "bge-small-en-v1.5",
          "input": ["hello world", "machine learning"]
        }
        ```

        Multimodal embedding (OpenAI-compatible):
        ```json
        {
          "model": "clip-vit-base-patch32",
          "input": [
            {"type": "text", "text": "a photo of a cat"},
            {"type": "image_url", "image_url": {"url": "data:image/png;base64,iVBORw0..."}}
          ]
        }
        ```
      operationId: generateEmbeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbedRequest"
      responses:
        "200":
          description: Embeddings generated successfully
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
                description: Binary-serialized embedding vectors
            application/json:
              schema:
                $ref: "#/components/schemas/EmbedResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /chunk:
    post:
      summary: Chunk text into smaller segments
      description: |
        Splits text into smaller chunks using semantic or fixed-size chunking models.

        ## Models

        ### Fixed Chunking (always available)
        - Simple token-based splitting with overlap
        - Use model="fixed"
        - Fast and deterministic

        ### ONNX Models
        - Semantic chunking based on content similarity
        - Models auto-discovered from `chunker_models_dir`
        - Falls back to fixed chunking if model fails

        ## Caching

        Results are cached in memory for 2 minutes. Cache key includes both config and text content.

        ## Example

        ```json
        {
          "text": "This is a long document...",
          "config": {
            "model": "fixed",
            "target_tokens": 500,
            "overlap_tokens": 50,
            "separator": "\n\n"
          }
        }
        ```
      operationId: chunkText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChunkRequest"
      responses:
        "200":
          description: Text chunked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChunkResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /rerank:
    post:
      summary: Rerank prompts by relevance
      description: |
        Re-scores pre-rendered text prompts based on relevance to a query using ONNX reranking models.

        ## Client Responsibilities

        The client must:
        1. Extract relevant fields from documents
        2. Render any templates
        3. Send pre-rendered text strings as `prompts`

        This design keeps Termite stateless and allows clients to customize rendering logic.

        ## Models

        - Models are auto-discovered from `reranker_models_dir`
        - Supports quantized models (`model_quantized.onnx`)
        - Automatically prefers quantized variants if available

        ## Example

        ```json
        {
          "model": "bge-reranker-v2-m3",
          "query": "machine learning applications",
          "prompts": [
            "Introduction to Machine Learning: This guide covers...",
            "Deep Learning Fundamentals: Neural networks are..."
          ]
        }
        ```

        For document-based reranking with field extraction, use the client-side
        `lib/reranking` package which handles rendering before calling this endpoint.
      operationId: rerankPrompts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RerankRequest"
      responses:
        "200":
          description: Prompts reranked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RerankResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Reranking service unavailable (no models configured)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /models:
    get:
      summary: List available models
      description: |
        Returns lists of available embedding, chunking, and reranking models.

        ## Embedders

        - ONNX models from `embedder_models_dir`
        - Quantized variants have `-i8-qt` suffix

        ## Chunkers

        - Always includes "fixed" (built-in)
        - Plus any ONNX models from `chunker_models_dir`

        ## Rerankers

        - ONNX models from `reranker_models_dir`
        - Empty if no models configured

        Models are discovered at service startup and cached.
      operationId: listModels
      responses:
        "200":
          description: Models retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelsResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...